Deprecate Option.IsZero and clarify in godoc
The only reason that the `IsZero` method exists, is to support the
`omitzero` tag in encoding libraries. Since our `Option` type is based
on Rust's `Option` enum, we're following the `Some`, `None` terminology.
Thus we cannot replace `IsNone` with `IsZero` just because the
`IsZero` name is now the de-facto standard in GoLang.

Closes #1